Extract BacktraceFormatters into a module.
This will be used by rspec/rspec-expectations#59.
