+++ lib/rspec/core/formatters/base_formatter.rb
@@ -115,29 +115,10 @@ module RSpec
           restore_sync_output
         end
 
-        def format_backtrace(backtrace, example)
-          return "" unless backtrace
-          return backtrace if example.metadata[:full_backtrace] == true
-
-          if at_exit_index = backtrace.index(RSpec::Core::Runner::AT_EXIT_HOOK_BACKTRACE_LINE)
-            backtrace = backtrace[0, at_exit_index]
-          end
-
-          cleansed = backtrace.map { |line| backtrace_line(line) }.compact
-          cleansed.empty? ? backtrace : cleansed
-        end
-
       protected
 
         def configuration
           RSpec.configuration
-        end
-
-        def backtrace_line(line)
-          return nil if configuration.cleaned_from_backtrace?(line)
-          RSpec::Core::Metadata::relative_path(line)
-        rescue SecurityError
-          nil
         end
 
         def read_failed_line(exception, example)
