+++ lib/rspec/core/formatters/helpers.rb
@@ -1,8 +1,34 @@
 module RSpec
   module Core
-    module Formatters
+    module BacktraceFormatter
+      extend self
 
+      def format_backtrace(backtrace, options = {})
+        return "" unless backtrace
+        return backtrace if options[:full_backtrace] == true
+
+        if at_exit_index = backtrace.index(RSpec::Core::Runner::AT_EXIT_HOOK_BACKTRACE_LINE)
+          backtrace = backtrace[0, at_exit_index]
+        end
+
+        cleansed = backtrace.map { |line| backtrace_line(line) }.compact
+        cleansed.empty? ? backtrace : cleansed
+      end
+
+    protected
+
+      def backtrace_line(line)
+        return nil if RSpec.configuration.cleaned_from_backtrace?(line)
+        RSpec::Core::Metadata::relative_path(line)
+      rescue SecurityError
+        nil
+      end
+    end
+
+    module Formatters
       module Helpers
+        include BacktraceFormatter
+
         SUB_SECOND_PRECISION = 5
         DEFAULT_PRECISION = 2
 
@@ -31,7 +57,6 @@ module RSpec
         def pluralize(count, string)
           "#{count} #{string}#{'s' unless count.to_f == 1}"
         end
-
       end
 
     end
