+++ lib/rspec/core/formatters/html_formatter.rb
@@ -86,7 +86,7 @@ module RSpec
           exception_details = if exception
             { 
               :message => exception.message, 
-              :backtrace => format_backtrace(exception.backtrace, example).join("\n")
+              :backtrace => format_backtrace(exception.backtrace, example.metadata).join("\n")
             }
           else
             false
